fix proc_pidpath and proc_regionfilename redundant memory allocation

+++ vmmap/src/lib.rs
@@ -27,13 +27,13 @@ pub use error::Error;
 pub trait VirtualMemoryRead {
     type Error: std::error::Error;
 
-    fn read_at(&self, offset: usize, buf: &mut [u8]) -> Result<usize, Self::Error>;
+    fn read_at(&self, offset: u64, buf: &mut [u8]) -> Result<usize, Self::Error>;
 }
 
 pub trait VirtualMemoryWrite {
     type Error: std::error::Error;
 
-    fn write_at(&self, offset: usize, buf: &[u8]) -> Result<(), Self::Error>;
+    fn write_at(&self, offset: u64, buf: &[u8]) -> Result<(), Self::Error>;
 }
 
 pub trait VirtualQuery {
